Reorganize scenario builder
