+++ airbyte-cdk/python/unit_tests/sources/file_based/scenarios/avro_scenarios.py
@@ -7,7 +7,7 @@ import decimal
 
 from unit_tests.sources.file_based.in_memory_files_source import TemporaryAvroFilesStreamReader
 from unit_tests.sources.file_based.scenarios.file_based_source_builder import FileBasedSourceBuilder
-from unit_tests.sources.file_based.scenarios.scenario_builder import TestScenarioBuilder
+from unit_tests.sources.scenario_based.scenario_builder import TestScenarioBuilder
 
 _single_avro_file = {
     "a.avro": {
