+++ airbyte-cdk/python/unit_tests/sources/file_based/scenarios/parquet_scenarios.py
@@ -9,7 +9,7 @@ import pyarrow as pa
 from airbyte_cdk.utils.traced_exception import AirbyteTracedException
 from unit_tests.sources.file_based.in_memory_files_source import TemporaryParquetFilesStreamReader
 from unit_tests.sources.file_based.scenarios.file_based_source_builder import FileBasedSourceBuilder
-from unit_tests.sources.file_based.scenarios.scenario_builder import TestScenarioBuilder
+from unit_tests.sources.scenario_based.scenario_builder import TestScenarioBuilder
 
 _single_parquet_file = {
     "a.parquet": {
